+++ opentelemetry-resource-detectors/src/aws.rs
@@ -1,0 +1,5 @@
+#[cfg(feature = "aws-lambda")]
+mod lambda;
+#[cfg(feature = "aws-lambda")]
+pub use lambda::LambdaResourceDetector;
+